Switch to ts and change up routes
