+++ calendar-component/src/admins/homepage.tsx
@@ -1,3 +1,5 @@
+import React from 'react';
+
 export default function Homepage() {
   return (
     <div>
