+++ calendar-component/src/auth/authHomepage.tsx
@@ -1,0 +1,8 @@
+import React from 'react'
+export default function AuthHomepage() {
+  return (
+    <div>
+      <h1>Auth</h1>
+    </div>
+  )
+}
