+++ calendar-component/src/auth/loginHomepage.tsx
@@ -1,0 +1,7 @@
+export default function LoginHomepage() {
+  return (
+    <div>
+      <h1>Login</h1>
+    </div>
+  )
+}
