+++ calendar-component/src/homepage.tsx
@@ -1,7 +1,9 @@
+import React from 'react';
+
 export default function Homepage() {
   return (
     <div>
-      <h1>Auth</h1>
+      <h1>Homepage</h1>
     </div>
-  )
+  );
 }
