--- calendar-component/src/index.js
@@ -1,38 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom/client';
-import reportWebVitals from './reportWebVitals';
-import { createBrowserRouter, Navigate, RouterProvider } from 'react-router-dom';
-import AdminHomepage from './admins/homepage';
-import UserHomepage from './users/homepage';
-import AuthHomepage from './auth/homepage';
-
-const router = createBrowserRouter([
-  {
-    path: "/admin",
-    element: <AdminHomepage />,
-  },
-  {
-    path: "/user",
-    element: <UserHomepage />,
-  },
-  {
-    path: "/auth",
-    element: <AuthHomepage />,
-  },
-  {
-    path: "/",
-    element: <Navigate to="/auth" />,
-  },
-]);
-
-const root = ReactDOM.createRoot(document.getElementById('root'));
-root.render(
-  <React.StrictMode>
-    <RouterProvider router={router} />
-  </React.StrictMode>
-);
-
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
-reportWebVitals();
